+++ dts-20200101/alibabacloud_dts20200101/client.py
@@ -140,6 +140,12 @@ class Client(OpenApiClient):
             query['DestCaCertificateOssUrl'] = request.dest_ca_certificate_oss_url
         if not UtilClient.is_unset(request.dest_ca_certificate_password):
             query['DestCaCertificatePassword'] = request.dest_ca_certificate_password
+        if not UtilClient.is_unset(request.dest_client_cert_oss_url):
+            query['DestClientCertOssUrl'] = request.dest_client_cert_oss_url
+        if not UtilClient.is_unset(request.dest_client_key_oss_url):
+            query['DestClientKeyOssUrl'] = request.dest_client_key_oss_url
+        if not UtilClient.is_unset(request.dest_client_password):
+            query['DestClientPassword'] = request.dest_client_password
         if not UtilClient.is_unset(request.destination_endpoint_data_base_name):
             query['DestinationEndpointDataBaseName'] = request.destination_endpoint_data_base_name
         if not UtilClient.is_unset(request.destination_endpoint_engine_name):
@@ -222,6 +228,12 @@ class Client(OpenApiClient):
             query['SrcCaCertificateOssUrl'] = request.src_ca_certificate_oss_url
         if not UtilClient.is_unset(request.src_ca_certificate_password):
             query['SrcCaCertificatePassword'] = request.src_ca_certificate_password
+        if not UtilClient.is_unset(request.src_client_cert_oss_url):
+            query['SrcClientCertOssUrl'] = request.src_client_cert_oss_url
+        if not UtilClient.is_unset(request.src_client_key_oss_url):
+            query['SrcClientKeyOssUrl'] = request.src_client_key_oss_url
+        if not UtilClient.is_unset(request.src_client_password):
+            query['SrcClientPassword'] = request.src_client_password
         if not UtilClient.is_unset(request.structure_initialization):
             query['StructureInitialization'] = request.structure_initialization
         if not UtilClient.is_unset(request.synchronization_direction):
@@ -287,6 +299,12 @@ class Client(OpenApiClient):
             query['DestCaCertificateOssUrl'] = request.dest_ca_certificate_oss_url
         if not UtilClient.is_unset(request.dest_ca_certificate_password):
             query['DestCaCertificatePassword'] = request.dest_ca_certificate_password
+        if not UtilClient.is_unset(request.dest_client_cert_oss_url):
+            query['DestClientCertOssUrl'] = request.dest_client_cert_oss_url
+        if not UtilClient.is_unset(request.dest_client_key_oss_url):
+            query['DestClientKeyOssUrl'] = request.dest_client_key_oss_url
+        if not UtilClient.is_unset(request.dest_client_password):
+            query['DestClientPassword'] = request.dest_client_password
         if not UtilClient.is_unset(request.destination_endpoint_data_base_name):
             query['DestinationEndpointDataBaseName'] = request.destination_endpoint_data_base_name
         if not UtilClient.is_unset(request.destination_endpoint_engine_name):
@@ -369,6 +387,12 @@ class Client(OpenApiClient):
             query['SrcCaCertificateOssUrl'] = request.src_ca_certificate_oss_url
         if not UtilClient.is_unset(request.src_ca_certificate_password):
             query['SrcCaCertificatePassword'] = request.src_ca_certificate_password
+        if not UtilClient.is_unset(request.src_client_cert_oss_url):
+            query['SrcClientCertOssUrl'] = request.src_client_cert_oss_url
+        if not UtilClient.is_unset(request.src_client_key_oss_url):
+            query['SrcClientKeyOssUrl'] = request.src_client_key_oss_url
+        if not UtilClient.is_unset(request.src_client_password):
+            query['SrcClientPassword'] = request.src_client_password
         if not UtilClient.is_unset(request.structure_initialization):
             query['StructureInitialization'] = request.structure_initialization
         if not UtilClient.is_unset(request.synchronization_direction):